Added a runner to benchmark shit

+++ CG_CodeVsZombies2/Utils/DistanceUtils.cs
@@ -4,7 +4,7 @@ namespace CG_CodeVsZombies2.Utils
 {
     public static class DistanceUtils
     {
-        public static double FastDistanceTo(ILocatable from, ILocatable to)
+        public static float FastDistanceTo(ILocatable from, ILocatable to)
         {
             int deltaX = to.X - from.X;
             int deltaY = to.Y - from.Y;
